pyproject.toml
[tool.poetry]
name = "bookstore"
version = "0.1.0"
description = "Bookstore API"
authors = ["Beatriz <beatrizleite_aragao@hotmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
django-rest-framework = "^0.1.0"
django-extensions = "^3.2.3"
django-debug-toolbar = "^4.2.0"
flake8 = "^6.1.0"
psycopg2-binary = "^2.9.9"


[tool.poetry.dev-dependencies]
pytest = "^7.4.2"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
